# Add SymbolFunc to precompile sympy expressions for faster resolution