--- a/cirq-core/cirq/protocols/json_test_data/spec.py
+++ b/cirq-core/cirq/protocols/json_test_data/spec.py
@@ -160,6 +160,7 @@ TestSpec = ModuleJsonTestSpec(
         'SupportsQasmWithArgsAndQubits',
         'SupportsTraceDistanceBound',
         'SupportsUnitary',
+        'SymbolFunc',
         # mypy types:
         'CIRCUIT_LIKE',
         'DURATION_LIKE',