--- a/cirq-core/cirq/study/symbol_func.py
+++ b/cirq-core/cirq/study/symbol_func.py
@@ -10,6 +10,7 @@ if TYPE_CHECKING:
 
 T = TypeVar('T')
 
+
 class SymbolFunc:
     """A "lambdified" symbolic expression that is faster for parameter resolution."""