--- a/cirq-core/cirq/study/symbol_func.py
+++ b/cirq-core/cirq/study/symbol_func.py
@@ -1,8 +1,23 @@
+# Copyright 2022 The Cirq Developers
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from typing import AbstractSet, TYPE_CHECKING, TypeVar, Union
 
 import sympy
 
-from cirq import protocols
+import cirq._compat as _compat
+import cirq.protocols as protocols
 
 if TYPE_CHECKING:
     import cirq
@@ -29,7 +44,7 @@ class SymbolFunc:
         self.func = sympy.lambdify(self.params, expr)
 
     def __repr__(self) -> str:
-        return f'SymbolFunc({self.expr!r})'
+        return f'cirq.SymbolFunc({_compat.proper_repr(self.expr)})'
 
     def _is_parameterized_(self) -> bool:
         return True