@maffoo 2022-03-02
I noticed in the docs for [`sympy.lambdify`](https://docs.sympy.org/latest/modules/utilities/lambdify.html?highlight=lambdify#sympy.utilities.lambdify.lambdify) that it uses `exec` internally, which is a security risk, so might have to forget about this. I still think that having some way to compile sympy expressions could help performance a lot, but we probably want something safer.