@maffoo 2022-03-02
Maybe we could add our own basic implementation of `lambdify` that would handle a subset of sympy expressions safely, roughly the same set of expressions that we already special case in `ParamResolver`: https://github.com/quantumlib/Cirq/blob/master/cirq-core/cirq/study/resolver.py#L139. We would basically want to produce an ast from the sympy expression, then compile it into a function. @dabacon, I know you have experience mucking with python asts from blqs. How hard do you think it would be to compile basic sympy expressions like this into functions?