@maffoo 2022-03-08
After mulling this over, I think using sympy.lambdify in the general case is probably not worth it. I'll take a stab at another approach to doing this.